Extract reusable proposal preview component (#881)
This change separates proposal preview rendering so the same content can
be reused in more places without coupling it to the full interactive
view.
